Added a simple Worker for Solidity compilation
