+++ js/src/redux/providers/worker.js
@@ -16,10 +16,16 @@
 
 import PromiseWorker from 'promise-worker';
 import runtime from 'serviceworker-webpack-plugin/lib/runtime';
+import WebWorker from 'worker-loader!~/webWorker.js';
 
 import { setWorker } from './workerActions';
 
-function getWorker () {
+// Setup the Service Worker
+setupServiceWorker()
+  .then(() => console.log('SW is setup'))
+  .catch((error) => console.error('SW error', error));
+
+function setupServiceWorker () {
   if (!('serviceWorker' in navigator)) {
     return Promise.reject('Service Worker is not available in your browser.');
   }
@@ -53,33 +59,32 @@ function getWorker () {
   });
 }
 
+function getWorker () {
+  try {
+    const worker = new PromiseWorker(new WebWorker());
+
+    return Promise.resolve(worker);
+  } catch (error) {
+    return Promise.reject(error);
+  }
+}
+
 export const setupWorker = (store) => {
   const { dispatch, getState } = store;
 
   const state = getState();
   const stateWorker = state.worker.worker;
 
-  if (stateWorker !== undefined && !(stateWorker && stateWorker._worker.state === 'redundant')) {
+  if (stateWorker !== undefined) {
     return;
   }
 
   getWorker()
     .then((worker) => {
-      if (worker) {
-        worker._worker.addEventListener('statechange', (event) => {
-          console.warn('SW state changed to', worker._worker.state);
-
-          // Re-install the new Worker
-          if (worker._worker.state === 'redundant') {
-            setupWorker(store);
-          }
-        });
-      }
-
       dispatch(setWorker(worker));
     })
     .catch((error) => {
-      console.error('sw', error);
+      console.error('setupWorker', error);
       dispatch(setWorker(null));
     });
 };
